requirements.txt
Django >= 5.1, <5.2
pillow
python-decouple
psycopg2-binary
selenium
urllib3
webdriver-manager
chromedriver-autoinstaller